+++ L06 Workspace/Acme-HandyWorker/src/main/java/controllers/EndorsementHandyWorkerController.java
@@ -16,6 +16,7 @@ import javax.validation.Valid;
 
 import org.joda.time.LocalDate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.util.Assert;
 import org.springframework.validation.BindingResult;
@@ -28,10 +29,12 @@ import security.LoginService;
 import services.CustomerService;
 import services.EndorsableService;
 import services.EndorsementService;
+import services.FixUpService;
 import services.HandyWorkerService;
 import domain.Customer;
 import domain.Endorsable;
 import domain.Endorsement;
+import domain.FixUp;
 import domain.HandyWorker;
 
 @Controller
@@ -46,6 +49,8 @@ public class EndorsementHandyWorkerController extends AbstractController {
 	private CustomerService		customerService;
 	@Autowired
 	private HandyWorkerService	handyWorkerService;
+	@Autowired
+	private FixUpService		fixUpService;
 
 
 	//	@Autowired
@@ -57,7 +62,7 @@ public class EndorsementHandyWorkerController extends AbstractController {
 		super();
 	}
 	@RequestMapping(value = "/create", method = RequestMethod.GET)
-	public ModelAndView create(@RequestParam("receiverId") final int receiverId) {
+	public ModelAndView create(@RequestParam(value = "receiverId", defaultValue = "-1") final int receiverId) {
 		ModelAndView result;
 
 		Endorsement endorsement;
@@ -66,40 +71,45 @@ public class EndorsementHandyWorkerController extends AbstractController {
 
 		final Integer userAccountId = LoginService.getPrincipal().getId();
 		Assert.notNull(userAccountId);
-		final Customer customer = this.customerService.getCustomerByUserAccountId(userAccountId);
 		final HandyWorker handyWorker = this.handyWorkerService.getHandyWorkerByUserAccountId(userAccountId);
-		if (customer != null) {
-			final HandyWorker handyWorker2 = this.handyWorkerService.findOne(receiverId);
-			Assert.isTrue(customer != null && handyWorker2 != null);
-			endorsement.setendorsableReceiver(handyWorker2);
-			endorsement.setEndorsableSender(customer);
-			endorsement.setMoment(LocalDate.now().toDate());
-			result = new ModelAndView("endorsement/handyWorker/edit");
+		final Customer customer = this.customerService.findOne(receiverId);
+		if (customer == null && handyWorker != null) {
+			final Collection<FixUp> fixUps = this.fixUpService.findAll();
+			final String language = LocaleContextHolder.getLocale().getDisplayLanguage();
+			final Collection<FixUp> myFixUps = this.fixUpService.findAllByHWLogger();
+
+			result = new ModelAndView("fixUp/handyWorker/list");
+			result.addObject("fixUps", fixUps);
+			result.addObject("myFixUps", myFixUps);
+			result.addObject("language", language);
+			result.addObject("requestURI", "fixUp/handyWorker/list.do");
 		} else {
-			final Customer customer2 = this.customerService.findOne(receiverId);
-			Assert.isTrue(customer2 != null && handyWorker != null);
-			endorsement.setendorsableReceiver(customer2);
+			Assert.isTrue(customer != null && handyWorker != null);
+			endorsement.setendorsableReceiver(customer);
 			endorsement.setEndorsableSender(handyWorker);
 			endorsement.setMoment(LocalDate.now().toDate());
+
 			result = new ModelAndView("endorsement/handyWorker/edit");
+
+			result.addObject("endorsement", endorsement);
 		}
-
-		result.addObject("endorsement", endorsement);
-
 		return result;
 	}
 
 	@RequestMapping(value = "/edit", method = RequestMethod.GET)
-	public ModelAndView edit(@RequestParam("endorsementId") final int endorsementId) {
+	public ModelAndView edit(@RequestParam(value = "endorsementId", defaultValue = "-1") final int endorsementId) {
 		ModelAndView result;
 
 		final Endorsement endorsement;
-		Assert.isTrue(this.endorsementService.findOne(endorsementId) != null);
 		endorsement = this.endorsementService.findOne(endorsementId);
+		if (endorsement == null || endorsement.getEndorsableSender().getUserAccount().getId() != LoginService.getPrincipal().getId())
+			result = new ModelAndView("redirect:show.do");
+		else {
+			Assert.isTrue(this.endorsementService.findOne(endorsementId) != null);
+			result = new ModelAndView("endorsement/handyWorker/edit");
 
-		result = new ModelAndView("endorsement/handyWorker/edit");
-
-		result.addObject("endorsement", endorsement);
+			result.addObject("endorsement", endorsement);
+		}
 
 		return result;
 	}
@@ -117,39 +127,48 @@ public class EndorsementHandyWorkerController extends AbstractController {
 				Assert.isTrue(endorsement != null, "endorsement.null");
 				final Endorsement saveEndorsement = this.endorsementService.save(endorsement);
 				result = new ModelAndView("redirect:show.do");
-				result.addObject("requestURI", "endorsement/handyWorker/show.do");
+				if (this.handyWorkerService.getHandyWorkerByUserAccountId(LoginService.getPrincipal().getId()) != null) {
+					result.addObject("deleteURL", "endorsement/handyWorker/delete.do?endorsementId");
+					result.addObject("requestURI", "endorsement/handyWorker/show.do");
+				} else {
+					result.addObject("deleteURL", "endorsement/handyWorker/delete.do?endorsementId");
+					result.addObject("requestURI", "endorsement/handyWorker/show.do");
+				}
+
 			} catch (final Throwable oops) {
 				System.out.println("El error es en endorsementController: ");
 				System.out.println(oops);
 				System.out.println(oops.getMessage());
 				System.out.println(oops.getLocalizedMessage());
 				System.out.println(binding);
-				result = new ModelAndView("endorsement/handyWorker/edit");
+				result = new ModelAndView("redirect:show.do");
 			}
 		return result;
 	}
 	@RequestMapping(value = "/delete", method = RequestMethod.GET)
-	public ModelAndView delete(@RequestParam("endorsementId") final int endorsementId) {
+	public ModelAndView delete(@RequestParam(value = "endorsementId", defaultValue = "-1") final int endorsementId) {
 		ModelAndView result;
 
 		final Endorsement endorsement = this.endorsementService.findOne(endorsementId);
 		System.out.println("endorsementId encontrado: " + endorsementId);
-		Assert.notNull(endorsementId, "endorsement.null");
-
-		try {
-			endorsement.getEndorsableSender().getEndorsementSender().remove(endorsement);
-			endorsement.getendorsableReceiver().getEndorsementReceiver().remove(endorsement);
-			final Endorsable saveEndorsableSender = this.endorsableService.save(endorsement.getendorsableReceiver());
-			final Endorsable savaEndrosableReceiver = this.endorsableService.save(endorsement.getendorsableReceiver());
-			this.endorsementService.delete(endorsement);
+		if (endorsement == null || endorsement.getEndorsableSender().getUserAccount().getId() != LoginService.getPrincipal().getId())
 			result = new ModelAndView("redirect:show.do");
-			result.addObject("endorsable", saveEndorsableSender);
-			result.addObject("requestURI", "endorsement/handyWorker/show");
-		} catch (final Throwable oops) {
-			System.out.println("Error al borrar endorsement desde hw: ");
-			System.out.println(oops);
-			result = new ModelAndView("redirect:show.do");
-		}
+		else
+			try {
+				Assert.notNull(endorsementId, "endorsement.null");
+				endorsement.getEndorsableSender().getEndorsementSender().remove(endorsement);
+				endorsement.getendorsableReceiver().getEndorsementReceiver().remove(endorsement);
+				final Endorsable saveEndorsableSender = this.endorsableService.save(endorsement.getendorsableReceiver());
+				final Endorsable savaEndrosableReceiver = this.endorsableService.save(endorsement.getendorsableReceiver());
+				this.endorsementService.delete(endorsement);
+				result = new ModelAndView("redirect:show.do");
+				result.addObject("endorsable", saveEndorsableSender);
+				result.addObject("requestURI", "endorsement/handyWorker/show");
+			} catch (final Throwable oops) {
+				System.out.println("Error al borrar endorsement desde hw: ");
+				System.out.println(oops);
+				result = new ModelAndView("redirect:show.do");
+			}
 		return result;
 	}
